requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & ML
openai==1.3.5
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Google Cloud (STT/TTS)
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1

# NLP
spacy==3.7.2

# Utilities
pydantic==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.25.1

# Development
pytest==7.4.3